Fix issue with forms in modals not working in iOS 11
